+++ src/config/events/index.ts
@@ -1,0 +1,35 @@
+export const events = [
+  {
+    id: '1',
+    heading: '#1 HackNITR',
+    title: 'HackNITR',
+    date: '29/02/2025',
+    description:
+      'Lorem ipsum dolor sit amet consectetur. Augue duis etiam pulvinar duis interdum accumsan tempor bibendum amet. Lorem ipsum dolor sit amet consectetur. Augue duis etiam pulvinar duis interdum accumsan tempor bibendum amet.Lorem ipsum dolor sit amet consectetur. Augue duis etiam pulvinar duis interdum accumsan tempor bibendum amet.',
+    imageUrl:
+      'https://res.cloudinary.com/dpmlrxlzr/image/upload/v1746378087/hn_50_1_ru18tc.png',
+    cardBgColor: 'bg-[#ECF3FE]',
+  },
+  {
+    id: '2',
+    heading: '#2 HackNITR',
+    title: 'HackNITR',
+    date: '29/02/2025',
+    description:
+      'Lorem ipsum dolor sit amet consectetur. Augue duis etiam pulvinar duis interdum accumsan tempor bibendum amet. Lorem ipsum dolor sit amet consectetur. Augue duis etiam pulvinar duis interdum accumsan tempor bibendum amet.Lorem ipsum dolor sit amet consectetur. Augue duis etiam pulvinar duis interdum accumsan tempor bibendum amet.',
+    imageUrl:
+      'https://res.cloudinary.com/dpmlrxlzr/image/upload/v1746378087/hn_50_1_ru18tc.png',
+    cardBgColor: 'bg-[#FDEDEB]',
+  },
+  {
+    id: '3',
+    heading: '#3 HackNITR',
+    title: 'HackNITR',
+    date: '29/02/2025',
+    description:
+      'Lorem ipsum dolor sit amet consectetur. Augue duis etiam pulvinar duis interdum accumsan tempor bibendum amet. Lorem ipsum dolor sit amet consectetur. Augue duis etiam pulvinar duis interdum accumsan tempor bibendum amet.Lorem ipsum dolor sit amet consectetur. Augue duis etiam pulvinar duis interdum accumsan tempor bibendum amet.',
+    imageUrl:
+      'https://res.cloudinary.com/dpmlrxlzr/image/upload/v1746378087/hn_50_1_ru18tc.png',
+    cardBgColor: 'bg-[#EBF7EE]',
+  },
+];
