+++ src/config/layout/index.tsx
@@ -1,0 +1,68 @@
+import React from 'react';
+export const pageContent = {
+  '/events': {
+    bgColor: '#E8F5EE',
+    title: 'Our Events',
+    description: (
+      <>
+        At Vero Et Accusamus Et
+        <br />
+        Lusto Odio Dignissimos The.
+      </>
+    ),
+    imageUrl:
+      'https://res.cloudinary.com/dpmlrxlzr/image/upload/v1746378087/hn_50_1_ru18tc.png',
+  },
+  '/about-us': {
+    bgColor: '#FFF4EC',
+    title: 'About Us',
+    description: (
+      <>
+        At Vero Et Accusamus Et
+        <br />
+        Lusto Odio Dignissimos The.
+      </>
+    ),
+    imageUrl:
+      'https://res.cloudinary.com/dpmlrxlzr/image/upload/v1746378087/hn_50_1_ru18tc.png',
+  },
+  '/projects': {
+    bgColor: '#F0F4FF',
+    title: 'Our Projects',
+    description: (
+      <>
+        At Vero Et Accusamus Et
+        <br />
+        Lusto Odio Dignissimos The.
+      </>
+    ),
+    imageUrl:
+      'https://res.cloudinary.com/dpmlrxlzr/image/upload/v1746378087/hn_50_1_ru18tc.png',
+  },
+  '/resources': {
+    bgColor: '#FFF4EC',
+    title: 'Resources',
+    description: (
+      <>
+        At Vero Et Accusamus Et
+        <br />
+        Lusto Odio Dignissimos The.
+      </>
+    ),
+    imageUrl:
+      'https://res.cloudinary.com/dpmlrxlzr/image/upload/v1746378087/hn_50_1_ru18tc.png',
+  },
+  '/team': {
+    bgColor: '#E8F5EE',
+    title: 'Our Team',
+    description: (
+      <>
+        At Vero Et Accusamus Et
+        <br />
+        Lusto Odio Dignissimos The.
+      </>
+    ),
+    imageUrl:
+      'https://res.cloudinary.com/dpmlrxlzr/image/upload/v1746378087/hn_50_1_ru18tc.png',
+  },
+};
